+++ lib/util/text/function_vars_test.go
@@ -24,6 +24,7 @@ import (
 	"fmt"
 	"os"
 	"strconv"
+	"strings"
 	"testing"
 	"time"
 
@@ -342,6 +343,21 @@ func TestFunctionSwapping(t *testing.T) {
 
 	if tThen.Sub(now).Seconds() > 5.0 {
 		t.Errorf("Timestamps too far out of sync: %v and %v", tThen, now)
+	}
+
+	now = time.Now()
+	tStamp = string(ReplaceFunctionVariables(nil, []byte("${!timestamp_utc}")))
+
+	tThen, err = time.Parse("Mon Jan 2 15:04:05 -0700 MST 2006", tStamp)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if tThen.Sub(now).Seconds() > 5.0 {
+		t.Errorf("Timestamps too far out of sync: %v and %v", tThen, now)
+	}
+	if !strings.Contains(tStamp, "UTC") {
+		t.Errorf("Non-UTC timezone detected: %v", tStamp)
 	}
 }